Refactor open-world layer management and improve type safety
- Replaced integer layer masks with a new `worldOpenLayerMask_t` type for better clarity and type safety in open-world functions.
- Updated functions to utilize the new layer mask definitions, enhancing readability and maintainability of the code.
- Simplified boolean return statements in various functions for cleaner code.
- Improved handling of sector loading and unloading by using defined masks instead of bitwise operations.

These changes enhance the robustness of the open-world feature set, making the codebase easier to understand and work with.

+++ src/navigation/nav_bsp_extract.h
@@ -6,6 +6,8 @@ BSP geometry extraction for Recast navmesh.
 */
 #pragma once
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -19,7 +21,7 @@ int    Nav_BSP_GetVertCount(void);
 int    Nav_BSP_GetTriCount(void);
 
 qboolean Nav_BSP_ExtractFromSectorMap( int cellX, int cellY, float sectorSize );
-qboolean Nav_BSP_ExtractFromSectorBuffer( const byte *buf, int length, int cellX, int cellY, float sectorSize );
+qboolean Nav_BSP_ExtractFromSectorBuffer( const byte *buf, size_t length, int cellX, int cellY, float sectorSize );
 
 #ifdef __cplusplus
 }
